backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Leo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = ">=4.2,<4.3"
djangorestframework = "^3.15.1"
python-decouple = "^3.8"
daphne = "^4.1.2"
channels = "^4.1.0"
channels-redis = "^4.2.0"
django-redis = "^5.4.0"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"
flake8 = "^7.0.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | migrations/
  | venv
  | _build
  | manage.py
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
skip = [".venv"]
